refactor(core): rename utils.h to time_utils.h
