+++ includes/acl/compression/impl/quantize_streams.h
@@ -27,8 +27,8 @@
 #include "acl/core/iallocator.h"
 #include "acl/core/impl/compiler_utils.h"
 #include "acl/core/error.h"
+#include "acl/core/time_utils.h"
 #include "acl/core/track_formats.h"
-#include "acl/core/utils.h"
 #include "acl/core/variable_bit_rates.h"
 #include "acl/math/quat_packing.h"
 #include "acl/math/vector4_packing.h"
